+++ NEWS.md
@@ -1,5 +1,16 @@
 # Package: sansSouci
 
+## Version: 0.1.3 [2016-01-20]
+
+* changed names of return values of 'getJointFWERThresholds' for consistency with the notation of the BNR paper.
+* added test script for unbalancedness.
+
+## Version: 0.1.2 [2016-01-07]
+
+* BUG FIX: the one-parameter family should not be updated at each step down!
+* Example fixed accordingly.
+* Implemented the "Oracle" version of step-down control.
+
 ## Version: 0.1.1 [2015-12-17]
 
 * added scripts to test step-down JFWER control